requirements.txt
fastapi[standard]~=0.114.0
fastapi-users[beanie]
uvicorn[standard]
pydantic>=2.8.0,<3.0.0
uvicorn==0.30.6
postmarker~=1.0
motor~=3.6.0
beanie~=1.26.0
bcrypt~=4.1.2
python-dotenv~=1.0.1
pymongo~=4.9.1
PyJWT~=2.8.0
# ollama~=0.3.3
# torch~=2.4.1
# transformers~=4.44.2